+++ NEWS.md
@@ -8,6 +8,9 @@
 
 * Added support for native execution on the macOS with the M1 ARM-based CPU. (#375)
 
+* Added threading support via `threads` argument for `$optimize()` and `$variational()` 
+  (was already available via `threads_per_chain` for `$sample()`). (#369)
+
 * Added `$sample_mpi()` for MCMC sampling with MPI. (#350)
 
 # cmdstanr 0.2.1